requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.5
filelock==3.12.4
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lit==17.0.1
llvmlite==0.41.0
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230918
packaging==23.1
pika==1.3.2
regex==2023.8.8
requests==2.31.0
sympy==1.12
tiktoken==0.3.3
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
